+++ src/GRA.Controllers/ViewModel/MissionControl/Schools/DistrictListViewModel.cs
@@ -1,9 +1,5 @@
 ﻿using GRA.Controllers.ViewModel.Shared;
-using System;
 using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace GRA.Controllers.ViewModel.MissionControl.Schools
 {
@@ -12,5 +8,6 @@ namespace GRA.Controllers.ViewModel.MissionControl.Schools
         public List<GRA.Domain.Model.SchoolDistrict> SchoolDistricts { get; set; }
         public PaginateViewModel PaginateModel { get; set; }
         public GRA.Domain.Model.SchoolDistrict District { get; set; }
+        public string Search { get; set; }
     }
 }