+++ src/GRA.Controllers/ViewModel/MissionControl/Schools/EnteredListViewModel.cs
@@ -1,9 +1,6 @@
 ﻿using GRA.Controllers.ViewModel.Shared;
 using Microsoft.AspNetCore.Mvc.Rendering;
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using System.ComponentModel;
 
 namespace GRA.Controllers.ViewModel.MissionControl.Schools
@@ -16,6 +13,7 @@ namespace GRA.Controllers.ViewModel.MissionControl.Schools
         public GRA.Domain.Model.School EnteredSchool { get; set; }
         [DisplayName("School")]
         public int SchoolId { get; set; }
+        public string Search { get; set; }
         public SelectList SchoolDistricts { get; set; }
         public SelectList SchoolTypes { get; set; }
         public int? CurrentPage { get; set; }