+++ src/GRA.Controllers/ViewModel/MissionControl/Schools/SchoolsListViewModel.cs
@@ -1,9 +1,6 @@
 ﻿using GRA.Controllers.ViewModel.Shared;
 using Microsoft.AspNetCore.Mvc.Rendering;
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace GRA.Controllers.ViewModel.MissionControl.Schools
 {
@@ -12,6 +9,7 @@ namespace GRA.Controllers.ViewModel.MissionControl.Schools
         public List<GRA.Domain.Model.School> Schools { get; set; }
         public PaginateViewModel PaginateModel { get; set; }
         public GRA.Domain.Model.School School { get; set; }
+        public string Search { get; set; }
         public SelectList SchoolDistricts { get; set; }
         public SelectList SchoolTypes { get; set; }
     }