+++ src/GRA.Controllers/ViewModel/MissionControl/Schools/TypeListViewModel.cs
@@ -1,8 +1,5 @@
 ﻿using GRA.Controllers.ViewModel.Shared;
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace GRA.Controllers.ViewModel.MissionControl.Schools
 {
@@ -11,5 +8,6 @@ namespace GRA.Controllers.ViewModel.MissionControl.Schools
         public List<GRA.Domain.Model.SchoolType> SchoolTypes { get; set; }
         public PaginateViewModel PaginateModel { get; set; }
         public GRA.Domain.Model.SchoolType Type { get; set; }
+        public string Search { get; set; }
     }
 }