+++ src/GRA.Domain.Repository/ISchoolDistrictRepository.cs
@@ -1,4 +1,5 @@
 ﻿using GRA.Domain.Model;
+using GRA.Domain.Model.Filters;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -7,8 +8,7 @@ namespace GRA.Domain.Repository
     public interface ISchoolDistrictRepository : IRepository<SchoolDistrict>
     {
         Task<ICollection<SchoolDistrict>> GetAllAsync(int siteId);
-        Task<DataWithCount<ICollection<SchoolDistrict>>> GetPaginatedListAsync(int siteId,
-            int skip,
-            int take);
+        Task<ICollection<SchoolDistrict>> PageAsync(BaseFilter filter);
+        Task<int> CountAsync(BaseFilter filter);
     }
 }