+++ src/GRA.Domain.Repository/ISchoolTypeRepository.cs
@@ -1,4 +1,5 @@
 ﻿using GRA.Domain.Model;
+using GRA.Domain.Model.Filters;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -7,8 +8,7 @@ namespace GRA.Domain.Repository
     public interface ISchoolTypeRepository : IRepository<SchoolType>
     {
         Task<ICollection<SchoolType>> GetAllAsync(int siteId, int? districtId = default(int?));
-        Task<DataWithCount<ICollection<SchoolType>>> GetPaginatedListAsync(int siteId,
-            int skip,
-            int take);
+        Task<ICollection<SchoolType>> PageAsync(BaseFilter filter);
+        Task<int> CountAsync(BaseFilter filter);
     }
 }